fix: README moved to root of package
