+++ packages/avilatek_bloc/lib/avilatek_bloc.dart
@@ -4,6 +4,6 @@ library avilatek_bloc;
 export 'src/pending_notifications/pending_notifications.dart';
 export 'src/permission_handler/permission_handler_bloc.dart';
 export 'src/permission_handler/permission_handler_listener.dart';
-export 'src/pick_and_upload_file/src/pick_and_upload_file_bloc.dart';
+export 'src/pick_and_upload_file/pick_and_upload_file_bloc.dart';
 export 'src/remote_data/remote_data_bloc.dart';
 export 'src/remote_data_paginated/paged_remote_data_bloc.dart';
