+++ packages/avilatek_bloc/lib/src/pick_and_upload_file/pick_and_upload_file_bloc.dart
@@ -1,12 +1,12 @@
 import 'dart:io';
 
-import 'package:avilatek_bloc/src/pick_and_upload_file/src/pick_and_upload_file_handler.dart';
+import 'package:avilatek_bloc/src/pick_and_upload_file/pick_and_upload_file_handler.dart';
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/foundation.dart';
 
-part 'package:avilatek_bloc/src/pick_and_upload_file/src/pick_and_upload_file_event.dart';
-part 'package:avilatek_bloc/src/pick_and_upload_file/src/pick_and_upload_file_state.dart';
+part 'package:avilatek_bloc/src/pick_and_upload_file/pick_and_upload_file_event.dart';
+part 'package:avilatek_bloc/src/pick_and_upload_file/pick_and_upload_file_state.dart';
 
 /// {@template upload_file_bloc}
 /// A [Bloc] that handles the picking and uploading of files.
